+++ CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/2.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.2] - 2026-07-??
+## [0.1.3] - 2026-07-27
+
+## Added
+
+- Added Spark tab
+  - Contains extra analysis completed with a Scala/Apache Spark batch pipeline
+  - Instructions to build Spark data available on the readme.md
+
+## [0.1.2] - 2026-07-27
 
 ## Added